course-project_assignmets/labs 1-4/requirements.txt
psycopg2-binary
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pydantic[email]>=2.7.0
python-dotenv>=1.0.0